Refactor: adjusting location of BILLING validations

+++ src/main/java/com/electricitybill/service/impl/CobrancaServiceImpl.java
@@ -3,10 +3,12 @@ package com.electricitybill.service.impl;
 import com.electricitybill.entity.CobrancaEntity;
 import com.electricitybill.repository.CobrancaRepository;
 import com.electricitybill.service.CobrancaService;
+import jakarta.persistence.NoResultException;
 import lombok.AllArgsConstructor;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Objects;
 
 @AllArgsConstructor
 public class CobrancaServiceImpl implements CobrancaService {
@@ -15,11 +17,17 @@ public class CobrancaServiceImpl implements CobrancaService {
 
     @Override
     public List<CobrancaEntity> findAll() {
+        if (cobrancaRepository.findAll().isEmpty()) {
+            throw new NoResultException("Search not found!");
+        }
         return cobrancaRepository.findAll();
     }
 
     @Override
     public CobrancaEntity findById(int id) {
+        if (Objects.isNull(cobrancaRepository.findById(id))) {
+            throw new NoResultException("Individual search not found!");
+        }
         return cobrancaRepository.findById(id);
     }
 
@@ -30,11 +38,17 @@ public class CobrancaServiceImpl implements CobrancaService {
 
     @Override
     public boolean update(int id, CobrancaEntity entity) throws SQLException {
+        if (Objects.isNull(cobrancaRepository.findById(id))) {
+            throw new NoResultException("Billing not found!");
+        }
         return cobrancaRepository.update(id, entity);
     }
 
     @Override
     public boolean delete(int id) throws SQLException {
+        if (Objects.isNull(cobrancaRepository.findById(id))) {
+            throw new NoResultException("Billing not found!");
+        }
         return cobrancaRepository.delete(id);
     }
 }
